requirements.txt
ics~=0.7
aiohttp~=3.7.4.post0
arrow~=0.14.7
PyYAML~=5.4.1
Telethon~=1.21.1